requirements.txt
cryptography==37.0.4
discord.py==2.0.0
mcipc==1.5.4
psutil==5.9.1
feedparser==6.0.10
jsons==1.6.3
omegaconf==2.2.2
colorama==0.4.5
requests==2.28.1
sshkeyboard==2.3.1